Deduplicate subscriptions
Currently every time someone subscribes, they will create a new
Reaction. The same thing happens when a subscription creates
subscriptions in it's own creation. The net result of this is that it is
possible to waste a lot of CPU time recalculating the same Reactions.

This patch adds a subscription cache. When a subscription is requested,
we check the cache. If it exists in cache then we return the cached
Reaction, if not then we create the subscription, add it to the cache,
and return it.

The tricky part is cache invalidation (isn't it always?). I'm not quite
sure how to remove Reactions that are no longer required.