middleware/debug: Log a single line if app-db not changed by a handler
middleware/debug logs a group to the console even if there has been no
change to app-db. This makes it harder to tell if you've made a mistake
in refactoring, as the same log message will be shown whether or not
app-db is changing.

This change will log a single line if app-db is not changed by a handler

Fixes #166