Introduce app-global-interceptors
Container and API to add application global interceptors to all events
without the need to define them on the `reg-event-` functions.

Adding intercetpors an the `reg-event-` functions is still possible
for fine grained control.

`re-frame.core` is refactored to use the new API to add its own global
interceptors and also the interceptor section is moved to the top of
the namespace.

The example todomvc app make use of the new API to add the `debug`
interceptor to all event handlers.

`app-global-interceptors` is a that map contains two vectors under the
keywords `:head`  and `:tail`.

Interceptors in the head vector come before the event interceptors, interceptors in
the tail vector come after. Event interceptors in this context are the ones defined in
`reg-event-` functions.