package.json file removed and replaced by bunch of lein aliases
Instead of using package.json which caused `lein deps` to fail. I
replaced it with several lein aliases which serve the same purpose.
For example `docs-build` builds the HTML version of docs in `./build/`
and `docs-serve` run a development server for docs on port `4000`.
For more information take a look at `project.clj`. NOTE: in order
to use the docs build system you need to install gitbook-cli utility.