Make subs only remove self from subscription cache when disposing
Adds a check for subscriptions to make sure that when they are dissocing
their cache key from the query cache that they are only removing
themselves. In certain cases when reloading/re-rendering the app,
subscriptions would be created and destroyed several times because each
version of subscription destroyed the following one.