Edit README.md
There are a lot of sentences in the first bit that start with "And" (unnecessarily). Removed the 3rd "and" start. AND the fourth.

Changed "I cannot stress enough what a big deal this is" to "The importance of this cannot be stressed enough" but I'd like a further change since I don't actually understand what "this" is fully in that context.

Changed "etcetera, boom, boom, boom" to "ad infinitum" 

There are far to many sentences starting with And. This should be changed in the future.