Change nil :fx value to warning instead of error
Middle ground fix for #737. I don't think silence on nil is OK esp in
the context of this error message being in place for some time. However
it needs to be changed to a warning so that users are not misled into
thinking that nil is unsupported - if you intended it, then it is fine.