Adds function to simplify creating new event types
Adds the function reg-event to re-frame.core which takes a handling interceptor
instead of a handler as its last argument. This makes it easier to create event
registering functions with all standard interceptors added to the chain,
even if changes are made to the standard interceptor chain.

reg-event-db, -fx and -ctx now wrap reg-event, passing their interceptor wrapped handlers.