[refactor] Consolidate db/path semantics
From one point of view, a namespace adds essential context to the
name. E.g., not just a path, but a path in terms of re-frame's app-db.

Arguably, the :re-frame.db concern doesn't depend on the present
concept of "store" or "history", but neither does :re-frame-path.

Since this is the only mention of :re-frame-path, I think :re-frame.db
is a more conventional point of context.