Fix flows that are both new & dead
A flow that's newly registered and live needs to calculate, regardless
if the inputs are dirty or not. This worked fine.

On the other hand, a flow that's new and dead would just stop being
new. Then later on, when it went live, it would not pass the dirty
check. This resulted in a flow not calculating when it should.

Now we're more explicit - :live and :new have different effects. Same
for :dead and :cleared.

Now, a flow stays new until it goes live, ensuring that it always
calculates the output when you need it.