Don't sort orphans
A graph like {:A [:B]} would crash the topsort, since :B isn't
actually a node. [:B] is an edge to nothing.

Now, these orphaned edges are simply removed. Since we expect users to
register and clear flows at runtime, it's natural that a flow might
depend on other flows which aren't registered. In that case, the input
will simply yield nil.