Fix cleared flows which depend on other cleared flows
Previously if a flow B depended on a flow A, and flow A was cleared,
flow B couldn't look up the old input of flow A, since flow A would no
longer be in the registry.

Now, we include the cleared flows when resolving inputs for flows, so
these cleared flows can still be considered when looking up old input
values.